Update transposeToitsown.c
Okay, so there was an issue or bug whatever. Actually  I was printing arr[j][i] which gives me the correct transpose, but the code does not change the main memory of the arr[i][j], because for the wrong loop condition, the loop changes the matrix twice, so it becomes same as it is, so now I have fixed this.

BTW why are you reading this ? Don't you have any work?